+++ src/GravitySimulation_v2/GravitySimulation_v2/settings.h
@@ -9,7 +9,7 @@ const int WIDTH = 1900;
 const int HEIGHT = 1100;
 
 // Timestep (speed of animation)
-const float TIMESTEP = 0.005f;
+const double TIMESTEP = 0.005;
 
 // Setup type
 const PlacementType PLACEMENT_TYPE = PlacementType::Circular;
@@ -20,10 +20,10 @@ const int N = 500;
 
 // Properties of particles
 const int RADIUS = 1;
-const float MASS = 2;
+const double MASS = 2;
 
 // Max force to expect per particle (for coloring purposes only)
-const float MAX_FORCES = 3.0f;
+const double MAX_FORCES = 3.0f;
 
 // Bounce back from edges of screen
 const bool HAS_BORDERS = true;
@@ -31,18 +31,18 @@ const bool HAS_BORDERS = true;
 // Trail settings
 const bool HAS_TRAIL = false;
 const int TRAIL_SIZE = 50;
-const float TRAIL_RADIUS = 0.5f;
+const double TRAIL_RADIUS = 0.5f;
 const sf::Color TRAIL_COLOR = sf::Color(40,40,40);
 
 // Collision settings (to fully turn off collision set both to false)
 const bool HAS_OVERLAPCHECK = true; // Turns on/off overlap avoidance
 const bool HAS_BOUNCEOFF = true; // Turns on/off bounce off when collide (conservation of momentum)
-const float COLLISION_THRESHOLD = 0.5f; // Amount of space left between particles
+const double COLLISION_THRESHOLD = 0.5f; // Amount of space left between particles
 
 // Small damping to avoid uncontrollable velocities (first dampens the angular velocity, then the final calculated velocity again)
 // THESE SETTINGS ARE CAREFULLY SET AND NOT ADVISED TO BE CHANGED
-const float DAMPING_COEFF = MASS * TIMESTEP / 10;
-const float DAMPING_DIST = 2.0f;
+const double DAMPING_COEFF = MASS * TIMESTEP / 10;
+const double DAMPING_DIST = 2.0f;
 
 // Collision optimization settings
 const bool VISUALIZE_SPATIAL_GRID = false; // draw grid
